move useWindowSize into separate file
