+++ src/components/Modal.js
@@ -7,11 +7,8 @@ const ModalStyles = styled.div`
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  transition: bottom 0.3s ease-out;
+  transition: all 0.3s ease-out;
   z-index: 10;
-  @media (max-width: 500px) {
-    width: 100%;
-  }
   > button {
     position: absolute;
     right: 6px;
@@ -20,6 +17,9 @@ const ModalStyles = styled.div`
   }
   &.Show {
     bottom: -50px;
+  }
+  @media (max-width: 500px) {
+    width: 100%;
   }
 `
 const OverlayStyles = styled.div`
@@ -42,9 +42,7 @@ const Modal = (props) => {
     <>
       {displayModal && 
       <ModalStyles className="Show">
-        <button className="Close minimal" onClick={() => setDisplayModal(!displayModal)}>
-        &times;
-        </button>
+        <button className="minimal" onClick={() => setDisplayModal(!displayModal)}>&times;</button>
         {children}
       </ModalStyles>}
       {displayModal && <OverlayStyles className='Show' onClick={() => setDisplayModal(!displayModal)} />}
